feat: turned MatrixAILintConfig and RawMatrixCfg into their own types and removed unnecessary interface. removed ignoring files console log
